+++ userspace/ksud/src/late_load.rs
@@ -35,6 +35,7 @@ fn dump_process_info(label: &str) {
 }
 
 pub fn run() -> Result<()> {
+    utils::daemonize(|| Ok(()))?;
     info!("late-load command triggered!");
     dump_process_info("late-load start");
